apps/cli/pyproject.toml
[project]
name = "clips-cli"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper",
    "numpy",
    "python-dotenv",
    "sentence-transformers",
    "google-genai",
    "tzdata",
    "groq",
    "assemblyai",
]
